error fix

+++ src/main/java/com/lxyw/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.lxyw.controller;
 
+import com.lxyw.entity.User;
 import com.lxyw.service.UserService;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,7 +17,7 @@ public class UserController {
     @RequestMapping("/showUser")
     @ResponseBody
     public User toIndex(HttpServletRequest request){
-        int userId = Integer.parseInt(request.getParameter("id"));
+        long userId = Long.parseLong(request.getParameter("id"));
         User user = this.userService.getUserById(userId);
         return user;
     }
@@ -24,7 +25,7 @@ public class UserController {
     @RequestMapping("/addUser")
     @ResponseBody
     public User addUser(HttpServletRequest request){
-        int userId = Integer.parseInt(request.getParameter("id"));
+        long userId = Long.parseLong(request.getParameter("id"));
         User user = this.userService.getUserById(userId);
         return user;
     }
